requirements.txt
keras==2.11.0
numpy==1.21
python-decouple==3.7
tensorflow-gpu==2.11.0
git+https://github.com/huggingface/transformers.git 
